abstract course searching to component
